requirements.txt
requests
beautifulsoup4
selenium
lxml
python-dotenv
